refactor: change package name && merge test file
